+++ release/ewoh-0.6.0-rc2/ewoh-spark-app/server/modules/onboarding/onboarding.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { ScaleModule } from '../scale/scale.module';
+import { OnboardingController } from './onboarding.controller';
+import { OnboardingService } from './onboarding.service';
+
+@Module({
+  imports: [ScaleModule],
+  controllers: [OnboardingController],
+  providers: [OnboardingService],
+  exports: [OnboardingService],
+})
+export class OnboardingModule {}